requirements.txt
selenium==4.10.0
pytest
flake8
python-dotenv
requests
pylint
pytest-reportportal
psutil